DERBY-5638 intermittent test failure in test_05_ClobNegative when running full largedata._Suite; LobLimitsTestjava.sql.SQLException: Table/View 'BLOBTBL' already exists in Schema 'APP'.
One of the tests run as part of large data suite is Derby5624Test and it has following in protected static Test baseSuite(String name) to set the lock timeout
       return new CleanDatabaseTestSetup(
               DatabasePropertyTestSetup.setLockTimeouts(suite, 2, 4))
       {

The test really doesn't need any special lock timeout setting and may be it is interferring with the locks timeouts that we see later on in subsequent suites in large data suite. This commit removes the unnecessay lock timeout setting from Derby5624Test.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1300459 13f79535-47bb-0310-9956-ffa450edef68