DERBY-5638 intermittent test failure in test_05_ClobNegative when running full largedata._Suite; LobLimitsTestjava.sql.SQLException: Table/View 'BLOBTBL' already exists in Schema 'APP'.
The issue has been that the CleanDatabaseTestSetup, after the last test fixture of embedded suite is done, tries to drop the tables but runs into lock timeout errors and hence it never finishes dropping the tables. None of these errors get reported anywhere by CleanDatabaseTestSetup and we simply move on to the next suite which in our case is network server running the large data tests. As part of CleanDatabaseTestSetup decorator for network server, we try to drop the existing objects in the database again before creating the new objects required by the new suite and the drop tables again run into lock timeouts. Now when the CleanDatabaseTestSetup creates the objects needed by network server suite, it runs into object already exists error. 

This commit will use singleUseDatabaseDecorator for the network server suite run of large data suite so a brand new database will be created and that way it will not run into left over database objects from the embedded run of large data tests. 




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1301109 13f79535-47bb-0310-9956-ffa450edef68