DERBY-5357 SQLJ.INSTALL_JAR shouldn't use identifier as file name
Since SQL identifiers can contain arbitrary characters, it is not safe
to use them as is as part of a file name. Trying ot map parts of the
name by excluding unsafe characters leads to a chance of name
collision. So, we have changed the naming altogether.

This patch, derby-5357-with-tests-4, changes the name (and location)
of the jar files stored in a database. The name is now based on UUID,
and no subdirectories under the directory "jar" are used: all
jar-files reside directly in the "jar" database directory, and the
name is of the form

<Derby uuid string>[.]jar[.]G[0-9]+

where <Derby uuid string> has the form hhhhhhhh-hhhh-hhhh-hhhh-hhhhhhhhhhhh
where h id a lower case hex digit, and the suffix ".G[0-9]+" is the
version number as before.

The format is changed on hard upgrade, cf tests in Changes10_9.

Also, dblook has been updated to cater for this change.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1302836 13f79535-47bb-0310-9956-ffa450edef68