DERBY-4275: Query executions fail when compressing a table using SYSCS_UTIL.SYSCS_COMPRESS_TABLE
backported changes #1142583 and #1160597 from trunk to 10.7 branch.

Move invalidation of dependent statements until the system tables have
been updated with information about the new conglomerates created by
compression or truncation. This is to prevent that statements
executing concurrently get recompiled too early and don't see the new
conglomerates (and therefore fail on subsequent executions because
they cannot find the old conglomerates). 

Fail in a controlled fashion (StandardException) if the conglomerate
disappears while binding the FromBaseTable. This used to cause a
NullPointerException. 



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.7@1304848 13f79535-47bb-0310-9956-ffa450edef68