DERBY-5489: getBinary() returns incorrect data after getObject() call on BLOB column
Added a test for the expected behavior of valid getters invoked on LOB columns. 
A restriction has been put in place to allow for only one getter invocation on
a given field, with the possible exception of getBytes and getString (iff
invoked as the first getter on the field).

Patch file: derby-5489-1b-test.diff


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1329186 13f79535-47bb-0310-9956-ffa450edef68