DERBY-5681 When a foreign key constraint on a table is dropped, the associated statistics row for the conglomerate is not removed
This problem happens because when two constraints share the same backing index, we conditionally dropped the statistics. Instead, this fix will make sure that the statistics are always dropped even if the underneath backing index is still valid(and hence won't be dropped and recreated) for other constraints. I ran derbyall and junit suite and they both ran fine with no errors.I have also dded few tests for the issue. 



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1329359 13f79535-47bb-0310-9956-ffa450edef68