DERBY-5565 Network Server should reject client connections that are not Derby  Network Client.
We now reject all PRDID's that do not start with DNC.
The protocol tests have been changed to use the DNC10090 PRDID instead of TST01000. This actually alters the code path where there is slightly different behavior in later derby client versions, so that we have to skip an additional dss on connect for the diagnostic information.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1336268 13f79535-47bb-0310-9956-ffa450edef68