Backporting 2 jiras to 10.4 with some manual changes as described below
DERBY-5681 When a foreign key constraint on a table is dropped, the associated statistics row for the conglomerate is not removed

DERBY-4834 Add helper class for working with index statistics in JUnit tests

DERBY-5681 added some tests to upgrade statistics procedure test but since that procedure doesn't exist in 10.4 and before, I added those tests to a new class, namely Derby5681Test.java. Also, because of the absence of update statisitcs procedure and DERBY-5702, we can't quite create a hanging statistics row in 10.4 and before. The test in 10.4 will just have a test to drop constraint(that shares a backing index with another constraint) and show that it doesn't break anything.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.4@1338905 13f79535-47bb-0310-9956-ffa450edef68