DERBY-5764: Make DatabaseMetaDataTest more robust wrt changes made by other tests
Use schema name in the queries to avoid "pollution" of the system tables from
other tests, specifically when run as part of the upgrade test. The schema name
is set to the user name, so to enable this feature wrap the test in a decorator
that changes the user.

Wrapped a ChangeUserDecorator around the tests from DatabaseMetaDataTest in the
upgrade tests.

Patch file: derby-5764-2a-specify_schema.diff



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1339240 13f79535-47bb-0310-9956-ffa450edef68