DERBY-5749 Implicit cast of variable length values, e.g. as arguments to stored methods and generated columns values, silently truncate if too long
Patches derby-5749b (stored procedures and functions) and
derby-5749-2b (generated columns).

Quote from releaseNote.html attached to the issue:

Summary of Change

SQL now does correct checking of the length of variable strings in
these two cases:

    Arguments to stored procedures and functions
    Values assigned to generated columns

Previously, if the actual value was longer than the datatype of the
argument or column to which it was assigned, Derby would silently
truncate the value and ignore the truncation. The SQL standard
requires a truncation exception be thrown.

Derby now throws an SQLException with SQL state 22001 in these cases.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1339281 13f79535-47bb-0310-9956-ffa450edef68