DERBY-5830: Make DoubleProperties.propertyNames() thread-safe
Don't store the property values in the intermediate Hashtable as they
are not needed. They may be null if the Properties instances are
modified after the recursive calls to Properties.propertyNames(), and
trying to store a null value in a Hashtable results in a
NullPointerException, causing issues such as DERBY-4269.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1353852 13f79535-47bb-0310-9956-ffa450edef68