DERBY-5833: Remove unused methods in NetCallableStatement class
This patch was contributed by Mohamed Nufail (nufail56 at gmail dot com)

This change removes the following unused methods from the
NetCallableStatement class:

- resetNetCallableStatement(NetAgent netAgent, NetConnection netConnection, String sql, Section section)
- resetNetCallableStatement(NetAgent netAgent, NetConnection netConnection, String sql, Section section, ColumnMetaData parameterMetaData, ColumnMetaData resultSetMetaData) 

I suspect that, of the two remaining overloads of the resetNetCallableStatement
method, the 3-argument variant could be made private, as it is called only
by the 6-argument variant. That might be a further improvement we could do
at some future time. For now, this change simply removes the two variants
that are wholly unused.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1356066 13f79535-47bb-0310-9956-ffa450edef68