DERBY-4279: Statement cache deadlock
Modified ResultColumnList.generateHolderMethod to only take a lock when
obtaining row template(s) if the data dictionary is in DDL_MODE. The lock that
used to be obtained was a CIS lock/latch.
The purpose of this change is to avoid the deadlocks that have been observed in
some scenarios when statements are being compiled or recompiled.

Patch contributed by Brett Wooldridge (brett dot wooldridge at gmail dot com).

Patch file: patch4279_2.txt


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1365661 13f79535-47bb-0310-9956-ffa450edef68