DERBY-5792: Make it possible to turn off encryption on an already encrypted database.
Added the new connection attribute "decryptDatabase", which will decrypt an
encrypted database given the correct boot password or key if set to true.
The request is ignored if the database has already been booted.
If authentication and authorization are both enabled, only the DBO is allowed
to decrypt the database.

Enabled the DecryptDatabaseTest, and added new tests to DboPowersTest and
CryptoCrashRecoveryTest.

A few high-level comments:
o EmbedConnection
Deals with connection attributes and DBO powers.

o RawStore
Deals with the meat of the feature; configuring cryptograhic operations (based
on connection attributes and service properties), modifying
service.properties, and extra crash recovery logic.

o RAFContainer
This is where we skip encrypting the data pages when writing the containers,
effectively decrypting the database.

Patch file: derby-5792-3a-decryption_feature.diff,
            derby-5792-4b-crash_and_dbo.diff



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1393993 13f79535-47bb-0310-9956-ffa450edef68