DERBY-5232 (Put a stern README file in log and seg0 directories to warn users of corrpution they will cause if they touch files there)
Adding a test case for default database creation with the checks for the existence of the 3 readme files. These readmes warn users against editing/deleting any of the files in the database directories. The location of the readme files are  
 1)at the db level directory, 
 2)in seg0 directory and 
 3)in the log directocy.
All the three readme files are named README_DONT_TOUCH_FILES.txt



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1409100 13f79535-47bb-0310-9956-ffa450edef68