DERBY-5894: NPE in OnlineBackupTest1 while backing up in stubFileToRemoveAfterCheckPoint
Stop using the shared run() method from RAFContainer.backupContainer(),
as it is not safe to use without synchronization (which is why the
NullPointerException happened).

Also, use doPrivileged() only to execute those parts of the code that
need to run with privileges, as most of the code called from
backupContainer() does not need privileges.

git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1422845 13f79535-47bb-0310-9956-ffa450edef68