DERBY-5945 Inappropriate error message when calling a procedure as a function (as if it had a return value)
Committed patch "DERBY-5945". It adds new usage checks for a procedure
called as a function and vice versa, two new eror messages,
LANG_PROC_USED_AS_FUNCTION (42Y03.S.3) and LANG_FUNCTION_USED_AS_PROC
(42Y03.S.4) as well as new test cases to LangProcedureTest.

The patch also does some cleanup of LangProcedureTest and activates to
be run as a standard part of the regressions in the lang suite.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1429197 13f79535-47bb-0310-9956-ffa450edef68