DERBY-5955 Prepare Derby to run with Compact Profiles (JEP 161)
Patch derby-5955-new-non-jndi-ds-02. This introduced the six (three
for embedded, three for client drivers) new reduced functionality data
sources that will work in the absence of JNDI. It adds a sanity test
which will run even on a full Java SE platform. Full tests to be
exercised when running in a constrained environment will follow in a
subsequent patch.




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1432380 13f79535-47bb-0310-9956-ffa450edef68