DERBY-5955 Prepare Derby to run with Compact Profiles (JEP 161)
Patch derby-5955-add-cp2-to-jars. This add a new attribute to the
manifest in the jar files to make sure the user will get a good error
message if Derby is attempted run with compact profile less than 2,
i.e. CP1:

MANIFEST.MF:
:
Profile: compact2 
:

Cf. this quite in http://openjdk.java.net/jeps/161:

"jar — The JAR-file manifest specification will be extended with a new
attribute which can be used to specify the minimum Profile required by
the code in a JAR file."



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1437995 13f79535-47bb-0310-9956-ffa450edef68