DERBY-5955 Prepare Derby to run with Compact Profiles (JEP 161)
Patch derby-5955-javadoc, which updates the javadoc with a single line
usage advice at the top of the javadoc for all the data sources, and
also adds description of the new data sources to the jdbc package's
javadoc. Removed ClientBaseDataSource from the public api, added
adding ClientDriver40 to the jdbc4 instead of ClientDriver.

git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1440264 13f79535-47bb-0310-9956-ffa450edef68