DERBY-5546 ResultSet#updateBigDecimal on a REAL column does not do underflow checking
Patch derby-5546-2. For both Real and Double, check for underflow. For
Double underflow is currently detected but only because we didn't fix
DERBY-3398 yet, so we introduce the same check now as for Real.  Once
DERBY-3398 it will no longer be redundant. The tests are still guarded
by a check for embedded until DERBY-5534 is fixed.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1447996 13f79535-47bb-0310-9956-ffa450edef68