DERBY-6045 (in list multi-probe by primary key not chosen on tables with >256 rows)
Splitting a junit fixture into two so that we have different tests for table with just primary key and table with primary and unique key. Additionally, got rid of the extra column in the table which is not part of any index to keep things simple and is not impacting the behavior of the queries in question.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1451683 13f79535-47bb-0310-9956-ffa450edef68