DERBY-6092: IOException when closing LOBFile at end of transaction
A bug in Java 5 may cause IOException if two threads attempt to
close the same RandomAccessFile at the same time.

Avoid this situation by removing the LOBFile from the list of open
files before closing it from LOBStreamControl's finalizer. Then the
cleanup code at the end of the transaction won't find the file and
try to close it if the finalizer is about to close it.

git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1466362 13f79535-47bb-0310-9956-ffa450edef68