DERBY-6122 -  InterruptResilienceTest fails with: junit.framework.ComparisonFailure: Database shutdown expected:<[08006]> but was:<[XBDA0]>
backported change #1466225 from trunk to 10.10 branch.

The core problem that led to this assert was that an attempt by the
test to shutdown the database timed out with:
Caused by: java.sql.SQLException: Login timeout exceeded.

The test had set a 10 second timeout assuming that was large enough such that
no timeout would be encountered, but failed a few times in public IBM lab
nightly regression runs.  Raising the number to 1000.




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.10@1466635 13f79535-47bb-0310-9956-ffa450edef68