DERBY-6144 nightly regression test failure, intermittent error : testStatisticsCorrectness(org.apache.derbyTesting.functionTests.tests.store.AutomaticIndexStatisticsTest)junit.framework.A
test was asserting that stats had to be created after "now".  I think the
nightly was getting a case where the time was the same.  DERBY-5046 fixed
a different part to the same test to check for greater than now, so implemented
that same fix at the offending line.  Also added some text to be printed with
the variable values if it happens again.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1467011 13f79535-47bb-0310-9956-ffa450edef68