DERBY-6163: Reduce visibility of methods in subclasses of PageBasicOperation
Make all abstract classes in the PageBasicOperation class hierarchy
package protected. Make all constructors in the class hieararchy
package protected, except the no-argument constructors as the
Formatable interface requires them to be public.

Also reduce visibility of the RawField class, which is only used by
StoredPage in the same package.

git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1467197 13f79535-47bb-0310-9956-ffa450edef68