DERBY-6179 Insert some guarding code into the first replication test to check port availability
Patch replscratch: checks the availability of the ports needed for the
replication test testReplication_Local_TestStraightReplication (the
first fixture) of the first test in the replication suite. This should
make it readily apparent in test log if there is such a
problem. Presently one needs to dig into the master or slaves' log
files to find that root cause.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1467722 13f79535-47bb-0310-9956-ffa450edef68