DERBY-6125 Code clean up in client driver.
derby-6125-rename-c which renames the name overloaded JDBC classes in
the client driver thus:

    am/<JDBCClassName>.java -> am/Client<JDBCClassName>.java

This allows import statements to be used for all JDBC and Derby
classes. The rest of the patch implements such usage instead of full
pathnames in usages. After this patch, the only overload that needs
explicit package path is "java.util.Date" which would otherwise
collide with java.sql.Date. In one instance, I let a full path remain
in order to avoid a compiler warning: the deprecated use of
StringBufferInputStream in ClientClob.java.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1471245 13f79535-47bb-0310-9956-ffa450edef68