DERBY-6169: Reduce visibility of classes and methods under impl/sql
Remove the CreateViewNode.getProviderInfo() method in which a bug (it
returns colInfos instead of providerInfos) was introduced in revision
1477476.

Also remove the single caller of that method, as it appears to be dead
code that was only needed for views that had been created with a
pre-Derby version of the engine. Since upgrade is only supported from
other Derby versions, that logic should not be needed anymore.

git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1479562 13f79535-47bb-0310-9956-ffa450edef68